Split primary_key gen as single file
